requirements.txt
asgiref==3.5.2
Django==4.0.6
django-crispy-forms==1.14.0
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
fontawesomefree==6.1.2
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.1